themix: verify client signature
`main.go`: add two flags, ck and sign. `ck` is the path of ecdsa private key, which is used for client signature verification. `sign` is a bool flag to decide whether to verify client signature or not.

`node.go`: pass `ck` and `sign` to proposer.

`propose.go`: verify client signature if proposer.sign is set.